Ensure that MockActionContext.executeAction always returns Promise to mimick real executeAction
